update course filter
